Update auth_views.py
Brought auth_views.py from main. Updated content so it would function in line with logic. Added some front-end and back-end feedback for logging in and out. Changed "form" calling in login and register functions to be more descriptive for easier code reading. Made logout action require user to be logged in.